# src/apiserver/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseOps V2 API",
    "version": "2.0.0",
    "description": "Enterprise platform API for PulseOps V2. Supports database management, authentication (JWT + HttpOnly cookies), module hot-dropping, and system configuration. All URLs are defined in urls.json — this spec references relative paths under the /api prefix.",
    "contact": {
      "name": "PulseOps Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative to host — works on any environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained via POST /auth/login. Expiry: 24h."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly cookie set automatically on login. Same JWT as BearerAuth."
      }
    }
  },
  "paths": {
...
    "/database/config": {
      "get": {
        "tags": ["Database"],
        "operationId": "getDbConfig",
        "summary": "Get database configuration",
        "description": "Returns the current database configuration from DatabaseConfig.json (password masked).",
        "responses": {
          "200": { "description": "Database configuration.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      },
      "put": {
        "tags": ["Database"],
        "operationId": "updateDbConfig",
        "summary": "Update database configuration",
        "description": "Persists the provided database configuration to DatabaseConfig.json on the API server's filesystem. Takes effect on next connection pool creation.",
        "requestBody": {
          "required": true,
          "description": "Database configuration to save.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatabaseConfig" } } }
        },
        "responses": {
          "200": { "description": "Config saved to DatabaseConfig.json.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "500": { "description": "File write failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/database/config/test": {
      "post": {
        "tags": ["Database"],
        "operationId": "testDbConfig",
        "summary": "Test database configuration",
        "description": "Tests PostgreSQL connectivity using the provided configuration. Creates a temporary connection pool, runs SELECT version(), and returns latency. Pool is destroyed after test.",
        "requestBody": {
          "required": true,
          "description": "Database connection parameters to test.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatabaseConfig" } } }
        },
        "responses": {
          "200": { "description": "Connection test result.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConnectionTestResult" } } } }
        }
      }
    },
...
    "/database/instance": {
      "post": {
        "tags": ["Database"],
        "operationId": "createDatabase",
        "summary": "Create the database instance",
        "description": "Connects to the 'postgres' system database and runs CREATE DATABASE for the target database name from config. Returns success if database already exists.",
        "responses": {
          "200": { "description": "Database created or already exists.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "500": { "description": "Database creation failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Database"],
        "operationId": "deleteDatabase",
        "summary": "Delete the database instance",
        "description": "DESTRUCTIVE: Drops the entire PostgreSQL database including all schemas, tables, and data. Requires JWT authentication. Terminates all active connections before dropping.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Database dropped.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Drop failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/database/schema": {
      "get": {
        "tags": ["Database"],
        "operationId": "getSchemaStatus",
        "summary": "Get schema status",
        "description": "Checks whether the database schema is initialized (tables exist) and whether default data has been seeded (admin user, core modules). Returns table list if schema exists.",
        "responses": {
          "200": { "description": "Schema status.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SchemaStatusResult" } } } }
        }
      },
      "post": {
        "tags": ["Database"],
        "operationId": "createSchema",
        "summary": "Create database schema",
        "description": "Creates the core schema and 4 tables inside a transaction: system_users (auth), system_config (JSONB key-value), system_modules (hot-drop registry), system_logs (structured logs). Rolls back on any failure.",
        "responses": {
          "200": { "description": "Schema and tables created.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SchemaCreateResult" } } } },
          "500": { "description": "Schema creation failed (rolled back).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Database"],
        "operationId": "deleteSchema",
        "summary": "Delete database schema",
        "description": "DESTRUCTIVE: Drops the entire schema CASCADE, removing all tables and data. Requires JWT authentication. The database itself is not dropped — only the schema.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Schema deleted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/database/schema/seed": {
      "post": {
        "tags": ["Database"],
        "operationId": "seedDefaultData",
        "summary": "Seed default data",
        "description": "Seeds the default admin user (from DefaultAdminUser.json with bcrypt-hashed password) and registers core modules (platform_admin, auth). Skips if data already exists.",
        "responses": {
          "200": { "description": "Default data loaded.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DefaultDataResult" } } } },
          "500": { "description": "Seeding failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Database"],
        "operationId": "cleanDefaultData",
        "summary": "Clean default data",
        "description": "Removes all seeded default data: deletes the default admin user and unregisters core modules from the database.",
        "responses": {
          "200": { "description": "Default data removed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/database/stats": {
      "get": {
        "tags": ["Database"],
        "operationId": "getDatabaseStats",
        "summary": "Get database stats",
        "description": "Returns table names, row counts, and disk sizes for all tables in the schema. Requires JWT authentication.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Database statistics.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatabaseStatsResult" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}